@friedriche
Played around with a lot of attempts at how to handle passing the Hero object around so that we can use the data.
I'm not completely sure but using a loader and .getController() I was able to access a hero object(and its methods) from another controller class.

Will need to reorganize and experiment more, but it's promising.