@friedriche
losing my entire mind lol

Modified Controller / Scene transitions so they are not "initialized" on each transition.
So between main and fightscenesetup, each has a hero object that is persistent now.

But I don't have access to the hero object in the other one, I think this is solvable (idk if the method i used for this before works in this new setup) but im not sure yet. it is late.

Also refactored some names (shouldn't be an issue but let me know if it is)
no more helloapplication or hello-view