@friedriche
#11
Testing herosheet manipulation.
Some issues with not being initialized, and being null. I haven't found a solution but it shouldn't be a real problem, we will simply have to have a "load" button to be pressed if the player wants to see everything updated the first visit to a hero sheet.
(same issue with names from the fightscenesetup)

Passed hero 2 and 3 into more scenes.