@friedriche
#3, #5, #8, #11, #13, #15, #18
Added hero objects to all current scenes.

Groundwork is now setup to start testing an "Active" hero that loads and unloads data of the player hero objects so that we can have a fight scene coded to only require the "Active" hero object which hopefully will be able to dynamically reference an active hero set through the inspect hero scene.