@friedriche
#3
#11
Lot of troubleshooting mostly completed.

So with our "static" scenes it seems that data/info doesn't update as much as I assume it would (require a reload/revist for some reason I don't fully understand)

But it seems to work if we add updates to the active hero on button clicks.
- Still an issue when we need data to reupdate such as after a fight, the herosheet will NOT update the character sheet unless the player clicks the load button for the hero.