@friedriche
Hah I am learning.

#3
#11
So we CAN pass handles from our controllers into other controllers which allows us to call methods of a controller that is not the active scene.
This means, I can call the load methods for the HeroSheetController from the FightSceneSetupController so the information WILL be up to date on entry.

This should work for every case.